Add DagMC::test_volume_boundary() to test whether on-surface ray enters or leaves volume
This function allows DagMC client to query whether a ray that originates on
the surface of the volume is entering or leaving that volume.  This is useful
in particle tracking when a particle changes direction at a surface and client
code wants to know whether to advance it to the next volume.  The function
uses a RayHistory object from ray_fire if available to avoid making any
OBB queries.

Add (simple and primitive) unit tests for this function.

Also add DagMC::RayHistory::size(), returning the number of entries in a history.
This is mostly for debugging purposes, though it may be useful for client code
to know if a history object is empty (size == 0).

git-svn-id: https://svn.mcs.anl.gov/repos/ITAPS/MOAB/trunk@5080 6162379e-bd28-0410-9a7d-b7f4dcfcad3c